Reorganize/revise comments for most files.
* Adds short description to all *.zsh files.
* Reorganize comments for clarity and conciseness.
* Removes a few unused commands

+++ osx/aliases.zsh
@@ -1,22 +1,28 @@
-# Toggle wifi 
-alias off='/usr/sbin/networksetup -setairportpower en1 off'     # wifi off
-alias on='/usr/sbin/networksetup -setairportpower en1 on'       # wifi on
+# set OSX-specific command aliases
 
-# Recursively delete `.DS_Store` files
-alias cleanup="find . -name '*.DS_Store' -type f -ls -delete"
+if [ $(uname -s) = "Darwin" ] ; then
 
-# Kill processes
-alias killmenu='killall SystemUIServer' 		# top menubar
-alias killdock='killall Dock'                   # Dock
-alias killfinder='killall Finder'               # Finder
+  # toggle wifi on/off
+  alias off='/usr/sbin/networksetup -setairportpower en1 off'
+  alias on='/usr/sbin/networksetup -setairportpower en1 on'
 
-# Empty the Trash
-alias emptytrash="rm -rfv ~/.Trash"
+  # recursively delete `.DS_Store` files
+  alias cleanup="find . -name '*.DS_Store' -type f -ls -delete"
 
-# Show/hide hidden files in Finder
-alias show="defaults write com.apple.Finder AppleShowAllFiles -bool true && killall Finder"
-alias hide="defaults write com.apple.Finder AppleShowAllFiles -bool false && killall Finder"
+  # kill processes
+  alias killmenu='killall SystemUIServer' # top menubar
+  alias killdock='killall Dock'           # Dock
+  alias killfinder='killall Finder'       # Finder
 
-# Hide/show all desktop icons (useful when presenting)
-alias hidedesktop="defaults write com.apple.finder CreateDesktop -bool false && killall Finder"
-alias showdesktop="defaults write com.apple.finder CreateDesktop -bool true && killall Finder"
+  # empty the Trash
+  alias emptytrash="rm -rfv ~/.Trash"
+
+  # show/hide hidden files in Finder
+  alias show="defaults write com.apple.Finder AppleShowAllFiles -bool true && killall Finder"
+  alias hide="defaults write com.apple.Finder AppleShowAllFiles -bool false && killall Finder"
+
+  # hide/show all desktop icons (useful when presenting)
+  alias hidedesktop="defaults write com.apple.finder CreateDesktop -bool false && killall Finder"
+  alias showdesktop="defaults write com.apple.finder CreateDesktop -bool true && killall Finder"
+
+fi
